+++ packages/order_repository/lib/src/treatment/models/models.dart
@@ -1,6 +1,5 @@
 export 'drug_dose_model.dart' show DrugDose, DrugDoseDescription;
 export 'item_order_status_codes_texts.dart'
     show ItemOrderStatusCode, ItemOrderStatusText, getStatusCodeViewModelText;
-export 'patient_model.dart' show Patient, getNameTextFromPatient;
 export 'patient_treatment_product_item_model.dart'
     show PatientTreatmentProductItem, ProductDescription;
