+++ packages/patient_repository/lib/patient_repository.dart
@@ -1,3 +1,5 @@
 library patient_repository;
 
+export 'src/models/models.dart'
+    show Patient, PatientBirthDateAustralian, getNameTextFromPatient;
 export 'src/patient_repository.dart' show PatientRepository;
