+++ packages/patient_repository/lib/src/models/models.dart
@@ -1,0 +1,3 @@
+export 'patient_birthdate_australian_extension.dart'
+    show PatientBirthDateAustralian;
+export 'patient_model.dart' show Patient;
