+++ packages/patient_repository/pubspec.yaml
@@ -2,14 +2,17 @@ name: patient_repository
 description: Patient repository for clinic's patients.
 version: 0.0.1
 author: s
+publish_to: none
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  i18n_l10n_australian:
+    path: "../i18n_l10n_australian"
 
 dev_dependencies:
   flutter_test:
